Wire device attestation into the session, behind a three-way mode
The plumbing for the verification core added earlier. After sign-on the
BOS session is challenged with a nonce; the answer is verified against
the account's manifest; a session that cannot prove a device is closed,
logged, or ignored depending on BENCO_DEVICE_AUTH.

Three modes and it defaults to LOG, which is deliberate. Enforcing from
the first deploy means a bug locks out every account that has a device --
all of them -- leaving the management API on its unix socket as the only
way back in. Run on log until real sessions are seen passing, then flip
to enforce. "If you do not support the protocol you do not get to talk"
is the right design; it is not a reason to discover in production
whether the implementation agrees.

Only the BOS session is challenged. Chat and ChatNav connections are
opened by an already-signed-on client, so challenging those would ask
the same device to prove itself repeatedly for nothing.

Enforcement lives in the server loop rather than the service, because
closing a connection is the loop's business and refusing to ANSWER is
not the same as refusing to SERVE.

The mode is checked before the service or the instance is touched, so a
handler with attestation off reaches neither -- which is also what keeps
the zero-valued Handler in the existing tests working. It did not at
first, and the existing ClientOnline test found it by segfaulting on a
nil instance.

A domain tag now separates an attestation from a room-message signature.
Without it the two constructions produce identical bytes when a room is
named after an account and carries the nonce as its text, so a room
signature could be replayed as device proof. Tests on both sides fail if
they ever line up again, and the exact context bytes are pinned here
because client and server build them independently.

+++ wire/benco_keydir.go
@@ -381,7 +381,7 @@ const BENCOAttestNonceLen = 32
 // SNAC_0xBE00_0x000A_BENCOKeyDirAttestChallenge asks a session to prove which
 // device it is.
 type SNAC_0xBE00_0x000A_BENCOKeyDirAttestChallenge struct {
-	Version uint8
+	Version uint16
 	Nonce   []byte `oscar:"len_prefix=uint16"`
 }
 
@@ -391,14 +391,14 @@ type SNAC_0xBE00_0x000A_BENCOKeyDirAttestChallenge struct {
 // right one out of the manifest instead of trying all of them; it is checked for
 // membership regardless, so sending somebody else's buys nothing.
 type SNAC_0xBE00_0x000B_BENCOKeyDirAttestResponse struct {
-	Version   uint8
+	Version   uint16
 	SignKey   BENCOKey
 	Signature []byte `oscar:"len_prefix=uint16"`
 }
 
 // SNAC_0xBE00_0x000C_BENCOKeyDirAttestReply reports the outcome.
 type SNAC_0xBE00_0x000C_BENCOKeyDirAttestReply struct {
-	Version uint8
+	Version uint16
 	// Accepted is 1 when the session is now attested.
 	Accepted uint8
 }
